Refactor sender callback to listener.
The idea is to have senders that can notify to a
set of listeners the result of a payload send.

Changes introduced are:
* Rename callback to listener to be more representative.
* The listener methods now receive also the payload to
  have a reference to the payload that caused the result.
* Remove the option from config because now this listeners
  are set in the sender and the notifier should not know
  anything regarding them, that ends up being decoupled.

Move to an approach where senders can notify of results through
listeners instead of a callback executed every time an event is sent.

Also, modified the previous sender callback interface to a sender
listener name, and added the payload send to the listener's methods.