Support whether data comes from an uncaught exception or not
There are certain use cases where people want to know if a set of data
comes from an uncaught exception or not. Add this information to the
data object and to the filter interface. This way we can use our
existing filter and transform infrastructure for dealing with this data.

Fixes #93