Provide convenience setter for JsonSerializer
Fixes #164

The default sender is built with a default json serializer. If you do
not specify a sender there is currently no way to customize the json
serializer. This introduces a method `jsonSerializer` on the config
builder to set this value on the default constructed sender. If a sender
is already provided then this json serializer is ignored.