Fix for a 403 error response from Nginx
It seems possible to access ROllbar's servers and get a 403 Forbidden response.
The code didn't handle this well at all since it always expects a json response
from the server.
This patch checks if the regex matcher successed and gracefully returns an error
response so you can actually determine what is going on.